refactor the Profile page to split the content in different components
