+++ packages/core/admin/admin/src/pages/ProfilePage/tests/index.test.js
@@ -10,7 +10,9 @@ import Theme from '../../../components/Theme';
 
 jest.mock('../../../components/LocalesProvider/useLocalesProvider', () => () => ({
   changeLocale() {},
-  localeNames: ['en'],
+  localeNames: {
+    en: 'English'
+  },
   messages: ['test'],
 }));
 
@@ -45,18 +47,18 @@ const App = (
 describe('ADMIN | Pages | Profile page', () => {
   beforeAll(() => server.listen());
 
-  // beforeEach(() => {
-  //   jest.clearAllMocks();
-  // });
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
 
-  // afterEach(() => {
-  //   server.resetHandlers();
-  // });
+  afterEach(() => {
+    server.resetHandlers();
+  });
 
-  // afterAll(() => {
-  //   jest.resetAllMocks();
-  //   server.close();
-  // });
+  afterAll(() => {
+    jest.resetAllMocks();
+    server.close();
+  });
 
   it('renders and matches the snapshot', async () => {
     const { container } = render(App);
